Fix re-execution of cupAndlex Mojos
Previously, CupMojo and LexMojo checked whether the output file was
newer than the input file, and if it was, they'd simply bail out.
But that skipped the `project.addCompileSourceRoot()` call that's
needed to make those output files available to the other modules,
meaning that builds would succeed the _first_ time, and then fail
every time after until `mvn clean` was run.

Now, right before the early return `project.addCompileSourceRoot()`
is called to ensure that the output files are available on every run.